+++ src/components/modal/DetailModal.jsx
@@ -1,10 +1,10 @@
-import React, { useEffect } from "react";
+import React from "react";
 // import { contents } from "../../mocks/contentsData";
 import { CgClose } from "react-icons/cg";
 import IconButton from "../UI/IconButton";
 import { FiMoreVertical } from "react-icons/fi";
 import PropTypes from "prop-types";
-// import escapeKeyDown from "../../util/escapeKeyDown";
+import useEscapeKeyDown from "../../hooks/useEscapeKeyDown";
 
 DetailModal.propTypes = {
   setOpenDetail: PropTypes.func,
@@ -13,31 +13,12 @@ DetailModal.propTypes = {
   openDetail: PropTypes.bool,
 };
 
-function DetailModal({ setOpenDetail, id, person, openDetail }) {
+function DetailModal({ setOpenDetail, id, person }) {
   const modalCloseHandler = () => {
     setOpenDetail(false);
   };
 
-  // useEffect(() => {
-  //   if (openDetail) escapeKeyDown(setOpenDetail, !openDetail);
-  // }, [setOpenDetail, openDetail]);
-
-  useEffect(() => {
-    const handleKeyDown = (event) => {
-      if (event.key === "Escape") {
-        setOpenDetail(false);
-        console.log("a");
-      }
-    };
-
-    // 컴포넌트가 마운트되면 이벤트 핸들러를 등록합니다.
-    document.addEventListener("keydown", handleKeyDown);
-
-    // 컴포넌트가 언마운트되면 이벤트 핸들러를 제거합니다.
-    return () => {
-      document.removeEventListener("keydown", handleKeyDown);
-    };
-  }, [setOpenDetail]);
+  useEscapeKeyDown(setOpenDetail);
 
   return (
     <div
